BugFix[1807385] Fixed the traceability report.
Recovered the SL_MInOutTraceReports callout removed in the cleanup project.
Show the results in the main page instead of in a popup window.
Fixed the links in the results.
Fixed the styles of the report.
Fixed the Product field in the filters to set the Product description when is being loaded a product_id.

--HG--
extra : convert_revision : svn%3A9509db7c-3d11-0410-ba00-cf6d01345df7/trunk%404973