Related to issue 0006283: Openbravo servlet classes have (static) members which are changed by post/get actions
HttpSecureAppServlet.classInfo is now not set in the service method but in the init one, thus is only initialized
once per servlet. In addition the name which is used by Jasper Reports is not calculated based on the language
anymore, it's maintained only for backwards compatibility. The REPORT_TITLE parameter is now set automatically to
Reports in the renderJR method.

--HG--
extra : convert_revision : svn%3A9509db7c-3d11-0410-ba00-cf6d01345df7/trunk%4012350