related to isse 9011.
Code review.

Translation for dataValues with sections was done more in xmlVectorValue.printStringBuffer() making more than one call to DataValue.printGenerated() method, this made it not to work with in sections with more than one data array (as GeneralLedjer report).

Now translation for sections is done once before calling print, when preprocessing templates.