Fixed bug 14125: Added multicurrency support for Payment Aging Balance report
Added a Currency field when launching the report. The amounts in the
report are automatically converted to the selected currency.
If no rate conversion is defined for any currency on a payment,
the system shows an error.

--HG--
extra : rebase_source : e07cd3faa1cb31d17d2fd65f568a306741bb58a2