Fixes issue 14161: Pareto report: ABC codes do not get assigned correctly
Root cause and solution:
The inaccuracy of floating point precision causes the issue as it computes one value in report and different value in update category. In order to fix it, the precision has been truncated to 6 digit. Hence, it is working fine as expected.

Impacts:
There is no impacts due to the change.