Fixes issue 14297: Is possible to create a document for an organization with a warehouse defined for another organization
Solution:
SE_Order_BPartner is the callout which is responsible for populating warehouse field. There, I just followed the same way of populating warehouse as it's used in ComboReloads294 (callout for Organization).

Impacts:
As SE_Order_BPartner is used in both Purchase Order and Sales Order windows, This modified functionality can also be seen in Sales Order too.