Fixed bug 11460: Unpost a document in a closed period
Added a check in the FACT_ACCT_RESET process that is run
only when the user manually tries to unpost a document. If the
period is closed, an error is shown.

However, the global 'Reset accounting' process will not show any error in
case it finds any document in a closed period.

--HG--
extra : rebase_source : e9f8a0f0428096566f07c4f75dd7924121d4b7b2