Fixes issue 15120: Wrong message when creating a debt payment
Root-cause:
C_DEBT_PAYMENT_TRG is calculating the un-invoiced order quantity by adding up all together.  Since we 1 and -1, the summation results 0 which means no order is pending to be invoiced. Hence, 'OrderCompletelyInvoiced' error message is displayed.

Solution:
It seems negative quantity needs to be carefully handled. When the difference between ordered quantity and invoiced quantity is calculated, we have to consider the absolute value of quantities irrespective of its sign.

Impacts:
None