fixed bug 13687: Sort tabs are less intrusive when updating
Previously sort tabs tried to reorder all the items with an interval
between positions of 10. The only exception to this was in case the
item was in a module not in development.

Now it tries to do as less changes as possible, keeping the elements
in the same position as they were in case they were already sorted even
if their interval is not 10. When positioning a new element between 2
existent ones, it is tried its possition not to affect the other elements
but to look for a sequenece in between, only in case no free sequences
are available sequences for other elements are recalculated.