Remove implicit dal-connection commit from UsageAudit call in HSAS. part of the HSAS service method does an insert into the ad_session_usage_audit table (if usage audit is enabled). That was done using the currently active dal-connection and has an explicit dal-flush/-commit after it. That impacts (at least) performance of callout-execution in FIC badly, as it does an implicit flush/commit after each callout execution.
This change moves the the insert into an new jdbc connection/transaction which
is committed directly.

Note: that does change behavior slightly as before at the end of the service
method, called just before the real servlet for each request, a dal-commit was
done and now this is reverted. However potentially affected only include HSAS
itself & the pluggable authentication managers.