Invoice post (dated on prepayment date if any)
Uses previously created payment (in/out) for a business partner when
completing an invoice.

It creates a new payment associated with the invoice with:
- Amount = 0
- Generated Credit = 0
- Used Credit = Invoice's grand total amount
The previously payment used-credit-amount is updated to the invoice's
grand total amount.

For finding the credit payment is mandatory that the generated credit amount is
equal to the invoice's grand total amount and the used credit is 0. So the whole
credit payment is used to pay the invoice.
Note: no multicurrency is allowed, so payment and invoice currency must be equal.

--HG--
extra : rebase_source : f910a9b9e3fd0035b4f1025a3b205b9661c71578