Related to issue 19805: Modify grid properties of String and Foreign Key fields
Include info about the length of ForeignKeyUIDefinition in the getGridFieldProperties method. Edit the info about the length of StringUIDefinition so it is taken from field.getDisplayedLength() istead of field.getColumn().getLength(). This length info is then used to select the field that will be automatically expanded in the grid (if any).

--HG--
extra : rebase_source : bdde524bfd7819722cab8f2df8cb1969423fc82b