Fixes issue 19805: Fix function that sorts an array by the length property
To decide which field should be automatically expanded in the field, the array that contains all the fields is sorted by the length property. Then the available field with the highest length will the one being auto expanded. The function used to sort the array was faulty, now it is fixed: it puts first in descending order the fields that have defined the length property, and after them it puts the fields that do not have that property defined.

--HG--
extra : rebase_source : 147149b513760d7254e065b4e270bf45b68a0560