Related to issue 19539: Improve DynamicExpressionParser test
The DynamicExpressionParserTest has been improved, so it can extend BaseTest instead of TestCase. This way, there is no need to duplicate the code modified in the transformValue function.
The class DynamicExpressionParser has been adapted to be able to be used in a BaseTest. Now, if weld can not be used the get the list of fields and auxiliary inputs, that info will be taken directly from the tab.

--HG--
extra : rebase_source : d9c75c4053a7c171514aa53f1d52f79d1ab258ca