Fixes issue 19843: Audit section fields show the correct info
The valueMap that was being used to get the name of a user out of its identifier was not being initialized correctly, thus sometimes in the audit s
ection the identifier of the user was beign shown, instead of its name. The new initialization has been added to the remeberValues function of ob-v
iew-form. It is called when the response of the datasource is received. The new code only affects the createdBy and updatedBy fields, and only when
 its valueMap is not properly initialized.

--HG--
extra : rebase_source : 96c738e8406da9284860bdaa67db1fbd476e8fb7