Fixes issue 19694: Wrong behaviour when introducing a date in grid
The function update value of ob-formitem-date.js was only meant to be called in the cellEditEnd in the grid view or when the item loses focus in the form view. The behaviour was wrong because the updateValue function was also being called while the field was being edited. A flag (_preventDateParsing) has been added to the grid to ensure that the updateValue function is only called when needed.

--HG--
extra : rebase_source : 4948faae994b7a28a4149e7d493f8183ecd5014b