Related to issue 19694: Date value is now always updated after losing focus
After applying the previous fix (changeset 15739), if the date field had a callback, its value was not being updated correctly if it had the focus and the tab key was pressed. In the previous fix, if the _preventDateParsing tag was true, the date was neither parsed nor updated. Now if the flag is active the date is not parsed, but the value is updated anyway.

--HG--
extra : rebase_source : bf02176602f9e3610d0e21dda14f7b348b416d84