Related to issue 19694: Date value is now always updated after losing focus
After applying the previous fix (changeset 15739), if the date field had a callback, its value was not being updated correctly if it had the focus and the tab key was pressed. In the previous fix, if the _preventDateParsing tag was true, the date was neither parsed nor updated. Now if the flag is active the date is not parsed, but the value is updated anyway.

--HG--
extra : transplant_source : %5C%B4%C4%14%DC%80u%BAO%01%80h%B9%9F%29%82%86%EFI%2C