Fixes issue 19984: String and textarea items are no longer being truncated
The same field (grid.field) was being used for limiting the size of the input items and for calculating the width of the column. Now two different fields are defined in the UIDefinitions: 'length' to limit the size of the input items and 'displaylength' to set the length of the columns.

The code that was used to define the length of the columns is no longer using the 'length' property, now it uses the 'displaylength' property. It has been changed in two places:
- In the call of the getDefaultColumnWidth function
- In the sort function that is used to select the column to be autoexpanded

--HG--
extra : transplant_source : %B3o%B3%FBKmL%0A-%CFP%08%F3%AB%AD%D5F%21%DC%FF