Related to issue 19984: ForeignKeyUIDefinition uses ad_field.displaylength for length
After the fix of issue 19984 (changeset 15784), all the UIDefinition classes that defined the grid length property, took that value from ad_column.length. It works on most cases, but it does not in the ForeignKeyUIDefinition class. If it is used in the ForeignKeyUIDefinition class, when a new row is entered in the grid, the foreign key fields are not initialized properly.
Now the ForeignKeyUIDefinition class initializes the length grid field with the value of ad_field.displaylength. Changing this does not have any repercusion on determining the maximum number of characters allowed for the field. This class represents foreign keys, so its value can not be entered manually.

--HG--
extra : transplant_source : u%7C%16%D4o%A0%8E%935%DBu%A8%CB%3A%A7%F6%BD%AB%DC%82