Fixes issue 19956: Selecters are not validated on change
The property validateOnChange of ob-selector-item has been set to false. This way, the validation of the selectors is only done when an option is selected or when the content of the selector is deleted.

This change not only fixes the issue but also improves the behaviour of selectors, because prior to the fix, if a mandatory selector had text entered but no option selected, the save button was enabled. Now the save button is not enabled until an option is selected.