Fixes issue 20023: Autosave is throwing a wrong error message.
The error message was shown because a tab that belonged to the same entity had been previously saved, and in that case all tabs belonging to that entity must be refreshed.
This was already being done to the child tabs of the tab being saved, but not to the same-level tabs.
Now, when a tab is saved, all the views of the window check if they have to be refreshed. They will be refreshed if:
- They belong to the same entity of the tab just saved and
- They are not in the list of excluded tab. This list prevents the refreshing of tabs that do not need to be refreshed (the tab being saved and its parent tab, if any)