Fixes issue 20029: Tab bar is not created if it is not needed.
The hasChildTabs attribute initialized in ob-view-tab.js is used to decide whether a new tabset has to be created, and whether a view has to include a maximize/minimize button.
Prior to this fix, this attribute was set to true if the tab had any child, regardless of the type of the tab (accounting, translation or other (always visible). When a tab had only one accounting tab and the ShowAcct session preferences was false, the tabset was created even though it was not going to contain any visible tab.
Now, if a tab has only accounting or translation tabs, the value of hasChildTabs will depend on the value of the ShowAcct and ShowTrl session properties.