Fixes issue 20071: Autosave with date field focus works both in grid and form view
The previous fix only achieved to fix the issue when the change of tab was done while the date field was focus in the form view. Now it also works if it is focused in grid view.

This solution works for date fields that have callouts and for date fields that do not.
Additional changes have been made to allow to save a date multiple times in a row.

--HG--
extra : transplant_source : %09%2Ck%FF%0At5%10%5CDA%C3L6%A5f%C2PjR