Related to issue 20071: Added unit test compatibility.
The changes made to fix the issue 20071 worked when the change of the date fields were done through user interaction. The unit tests call directly the setValue method of the dateTextField property of the date, scenario which had not been considered, and the date field was not being flagged as changed. This has been fixed by overwriting the setValue method of the dateTextField of the date field.

--HG--
extra : rebase_source : 789dbdf0753e574dd8afa4d70b4b8ed9810e989b