Fixes issue 20023: Autosave now works in tabs childless tabs
There was an error in the previous fix that prevented the autosave from working in tabs that had no child tabs. This error (a check that the childTabSet of the view was not null) has been fixed.

--HG--
extra : rebase_source : adbda4932d195f0496de5e607e1c5a95d57697e0