Fixes issue 12910: RTL and LTR are generated when neede
Prior to the fix, if the user deleted the WebContent folder and executed smartbuild, the WebContent/web/skins/rtl and WebContent/web/skins/ltr folders were not generated unless the property changes.in.md5.sk
ins was true.

Now, both folders are regenerated when any of them are missing when executing the smartbuild.