Fixes issue 20147: A new record is created in the grid when pushing the enter key
A change (changeset 2e1947ca8bbc) introduced for trying to fix issue 19830 prevented a new line from being created when the user pushed the enter key while editing the last reco
rd in the grid. This change has been reverted, and the functionality is working again.

This change does not modify the status of issue 19830, because the changeset that has been reverted only solved part of the problem.