Fixes issue: 20187: Remove button is displayed correctly in Payment In
In the fix of issue 20035, the AddOrderOrInvoice.html file was modified in order to make it compatible with Internet Explorer. That change broke the creation of the remove buttons in the window that is opened when the Add Details button is pushed in the Payment In window. Now, instead of using appendChild(document.createTextNode(text)) to modify the innerHTML property of DOM  elementes, a div element is created and appended.

--HG--
extra : transplant_source : J%28%97C%943%90%8C%E9%E0%FE%EB%ED%87%5E%8B5%F8%7E%8E