Fixes issue 20241: It is possible to move between records while editing in grid.
In the changeset 2e1947ca8bbc, related to issue 19830, a change was done that had a side effect that prevented the user from beign able to move between records while editing the grid when the records were modified.

This change has been reverted and it does not have any consequence in issue 19830, because when the backout was done the issue was still not fixed.

--HG--
extra : transplant_source : g%C0%F9%BEM%FF9%91%0C%11%FC%D8e2%3F5ka%EF%F5