Fixes issue 20286: Quick launch combo shows all options
In the getPickListFilterCriteria function of ob-quick-launch.js, the function getValue() was beign used as the criteria. If no option has been previously selected in the combo, getValue() stores the text entered in the combo, otherwise it stores the title of the last window opened using the combo.
Now, the function getDisplayValue() is used for the criteria. This function retrieves the text entered in the combo, and is reset each time the combo is opened.