Related to issue 20192: adding a parameter to setFieldFormProperties
Before the fix, when a field is not shown in form view but displayed in grid view, the property editorProperties.visible is changed by setFieldFormProperties function. This value in the property doesn't allow to the field to get the focus in grid view. To solve it, a second parameter has been added to the function and setted as true in the affected call. If the function is called without setting the new argument its behaviour will be the same as always.

--HG--
extra : rebase_source : 99aed576f05b69f22dd282e2e1e156e67a2760e2