Fixes issue 20345: Code review changes applied
Now the field used to retrieve the language of the core module is ad_module_id (0) instead of name ('Core', it may vary in the future). Instead of executing the consult every time the language is needed, it is executed once and the returned value is saved in a variable.
The year in the license has been updated.