Fixes issue 20530: AccessibleEntities does not contain inaccessible entities
OB.AccessibleEntities is used to check if there should be a link in a title field. If a user does not have access to the window that will be opened when the link is clicked, that entity should not be contained in AccessibleEntities.

The initialization of AccessibleEntities has been modified, so if the window related to an entity is not accessible by the current role, it won't be contained in the list.