Fixes issue 20507: YesNo filter fields have been fixed.
There was a problem with the initialization of the YesNo fields in smartcliente. They were being initialized with 'Yes' or 'No' instead of with true and false, which are the possible expected values.

The getCriterion function has been overwritten, so it detects if the value is faulty and returns a proper one.