Fixes issue 20537: Wrong validation is not launched
When a row was saved, the last focused item was being stored. When that same line was opened to be edited, the validation was launched before fetching the row values, showing an error border on the last focused item. When the row values were fetched, the border went back to normal.

Now, when the form is hidden, it set to null the last focused item, preventing the wrong validation from beig launched.