Fixes issue 20874: Prevented exception when pushing the down arrow key
In the fix of issue 17910 some logic was introduced that checked the id of the last record of the grid. If that record was yet to be loaded, the exception 'Cannot read property from undefined' was raised.

Now, before executing that statement, it is checked if the last grid record is defined.

--HG--
extra : transplant_source : z%DA%3BA%A8z%90%D3%B3%F3%0E%E35%9C%92%5E%13%DB%15-