Pending receipts modal window is closed when selecting the current receipt
Before this fix, when the current receipt was selected in the pending receipts modal window, the modal window did not close.

This has been fixed by preventing the load of an order if it is the same as the current one.

--HG--
extra : rebase_source : 262b5f22923751c59af291066e7e9c6f95ba2cb9