Document number initialization modified
There have been some changes in the way the Document No of the orders is initialized and stored. Now the server provides the sequence number of the last processed order, and the document number prefix of the pos instance.

In terminal.js, upon receiving the pos initial data, it is checked if there are pending orders. If there are no pending orders, the sequence number received from the server is stored in the database. If there are pending orders, its sequence number is obtained and compared to the one received from the server. The maximum of them is stored in the database.

A new model has been defined to represent DocumentSequence, in order to be able to store it in the database.

Changes by files:
- document-sequence.js: A new Backbone model has been created in order to be able to store de document sequences in the database
- POSUtils.java: A new function has been added: getLastDocumentNumberForPOS. Given a POS search key, it gets the maximum sequence number of all its processed orders.
- Terminal.java: Now it provides to the POS its maximum sequence number and its document number prefix.
- index.jps: The document-sequence.js file has been added.
- order.js: Instead of using localStorage to retrieve the next document sequence, it gets it from a global variable OB.POS.modelterminal.get('documentsequence'). When a new order is added, this number is incremented, and both its global variable and its representation in the database are updated.
- terminal.js: Upon receiving the initial pos data from the server, it initializes the order sequence number in the database and in the global variable.
- pointofsale.js: The DocumentSequence model has been added to the array OB.Data['retail.pointofsale']

--HG--
extra : rebase_source : d140883c7c17c490f0de40b96edfdadea7b75440