Fixes issues 21066, 21094: Undo button appears when needed
In the fix of issue 21003, a change was introduced to avoid saving the undo property of the orders in the database. This was done in a way made the undo button to dissapear after an order was saved in the database. This has been changed, and now when the order is saved the undo button does not dissapear.

--HG--
extra : rebase_source : 06387a6486ef68d00941d5ee62829300b1201203