Fixes issue 21049: Callout messages are shown even if saving row
Callout messages will be shown always, regardless of if the callout has been called upon tabbing out of the field, saving its row or moving to another row.

To prevent the autosave from hiding the messagebar, now it will do it only if the message bar is showing an error message (as it was supposed to work, looking at the comments).