Fixes issue 20850: Alert hour is displayed correctly
The alert hour was not being saved correctly because of the way sqlc.java generates processAlertData.java. Even though the column in the database is a timestamp, in the select and in the insert functions the date format was being used, instead of the date time format.

It was not possible to fix it modifying the xsql file, and it was very risky to modify Sqlc.java, so now ProcessAlerts.java use jdbc to select and insert the alerts.

--HG--
extra : rebase_source : 55b0c79b776cc27838cdb26e9dbf4553c6982230