Fixes issue 21216: Now the proper grid buttons are shown
The toggleProgressIcon function was given for granted that when it was called the Show in form / Edit in grid buttons were visible. However, if the user modifies a field with a callout and presses the down arrow key, this function will be called with the Cancel / Save button visible. Now, it can be called in either of both states.