Fixes issue 20722: Selected record does not dissapear after button processing
When the grid is refreshed due to the execution of a process, the selected record does not dissapear, even if it does not comply with the filter criteria. But, if the user clicks on the Refresh Grid button, the only records that will be shown will be those that match the filter.