Fixes issue 20386: Filtering fields with multiple identifier now works
Prior to this fix, if a field identifier was composed by the concatenation of several fields and if its reference was Table, in certain conditions it was not possible to use this field in the filter. This happened because in the server the field used to filter was the first one used in the identifier, when the filter should have been done on the displayed field.

--HG--
extra : rebase_source : 96135fff9bb6bd3ff98cba6fbf8cb617cbe739fa