Fixes issue 21229: Parent property of tabs has been fixed
It was not possible to create new records in child tabs whose table was the same as its parent tab. This had not been a problem so far because all these tabs
were read only or single row, and in both cases it was not possible to add a new record.

The bug was located in the getParentProperty method of the OBViewTab class. Now it is taken into account the UIPattern of the tab, so if the tab is not read o
nly nor single row, it will be possible to add new lines.