Fixes issue 21306: Delete record shortcut improved
So far, the only check done when deleting a record with the shortcut Supr was to count the number of selected records, and not to check if the tab is read only, is the table is deleteable, etc.

Now, before deleting a record with the shortcut, the status of the delete button (where the logic to see if the removal of records is anabled) is checked. If the delete button is not enabled, it will not be possible to delete the record with the shortcut.

--HG--
extra : rebase_source : 0b2187c1a7aa7309e88eee92b3290b6d7258035c