Fixes issue 20564: Previous lines not shown when creating a new record
When a new record was created using the icon in Recent Views, if there was only one record in that window, it would be automatically selected, and its lines would be shown under the empty header tab.

This has been fixed by preventing this automatical selection, if the current record is new.