Fixes issue 21361: Data can be saved in local server without internet access
Before this fix, it was not possible to save data without being connected, because of this code in the saveData function:

        if (isc.Offline && isc.Offline.isOffline() && !this.dataSource.clientOnly) {
            isc.warn(this.offlineSaveMessage);
            return;
        }

Offline.isOffline() is defined as follows:

    isOffline : function () {
        if (this.explicitOffline !== null) return this.explicitOffline;
        var offline = window.navigator.onLine ? false : true;
        return offline;
    }

So, by setting Offline.explicitOffline to false, it is possible to go on with the saving process even if the browser detects there is no internet connectivity.

This value is set just before the save, and restored to its original value right after the saving process is completed.

--HG--
extra : rebase_source : 5fa49570c1ebf6fd71062a48285aaf52660d0621