Related to issue 21169: Change logic to prevent wrong protocol detection
The logic to detect the web protocols has been changed to prevent false positives. Before the fix, if the user defined a URL like htmlcodetutorial.com, the web protocol would not have been added because the URL starts with html.

Now this has changed, and supports http, https and ftp protocols.