Fixes issue 21375: Field is updated even if not visible in grid
When a row is being edited in the grid, if there is a callout and the changed column is not displayed in the grid, its value was not being updated.

This has been fixed by modifying the setItemValue function of ob-view-form, so that it now takes into account that a field not being shown in the grid might be updated.