Fixes issue 21352: Changing the shown fields while editing the grid is possible
Due to a change in smartclient, it was not possible to hide or show a field in the grid while editing a row. This happened because when a field was shown or hided, the edit values of the record were being deleted. This edit values were being used to retrieve all the record values in order to send them the the FIC. This deleted values were not being restored ever.

Now, to fix this, the edit values of the grid are saved when hiding or showing fields. If there is call to doChangeFICCall caused by this, the saved edit values will be used:

record = isc.addProperties({}, this.viewGrid.getRecord(rowNum), this.viewGrid.getEditValues(rowNum), this.viewGrid._savedEditValues);

This was only reproducible while editing new rows because saved rows had all the properties in the own record (this.viewGrid.getRecord()), so it did not matter that they were being deleted from the edit values.