Fixes issue 21568: Using the Product selector do not cause log errors
The fix used to detect unknown properties was applied twice: to fields that properties that contained the field separator and to fields that did not. It should only have been applied to the fields that do not contain the field separator (if it contains a field separator, we can be confident the property is sent from the Openbravo client).

--HG--
extra : transplant_source : j%B1%C8%F8%F0%09%C6%9FB%A5c%8Fr%E8%F5%7B%9Er6%06