Fixes issue 21470: Selectors with 'startsWith' text match style fixed
Before the fix, when some text was entered in a selector control, all options that contained that text were shown. This happened even if the suggestion text match style was set to 'Starts with', where in that case only the options that contains the text as a prefix should have been shown.