Fixes issue 21419: TAB key in drop down list does not skip a field
When an option is picked in a drop down list using the tab key, one field used to be skipped, being focused the wrong one. This happened because when the tab key is pressed the focus is automatically moved to the next item, as well as when an option was picked in a drop down. Now, when picking an option in a drop down, the focus will be only placed in the next field if the key that triggered the event is not the tab key.