Fixes issue 21657: It is possible to save views that have ids in its fields
If a window that an id in its fields, it was not possible to save new records because the ID was a mandatory field without value, so the save button was not enabled. Now the allRequiredFieldsSet function has been modified, so it does not take into account id fields, because its value will be set in dal before inserting the new row in the database.

--HG--
extra : rebase_source : 3bfb805f533f1bd561e3b4c3057c27b3fa071901