Fixes issue 21654: Selecting a blank option enables the save toolbar buttons
In the pickValue of the ob-formitem-list file, the _value attribute was set to '' to always trigger the update. This did not work because then selecting a blank option would not trigger the update.

Now, the _value attribute is set to a random number.

--HG--
extra : rebase_source : 016c5bb5d9042b3e708f6b1991f2d17b98548752