Fixes issue 21676: Tabbing out right after filtering in the grid now works
The problem was that if the user entered some text in the filter (i.e. 'Ce') and right after and before the datasource response was received he moved to the next field, the entered text was deleted and the datasource filter response was not applied.

This has been fixed by reentering the deleted text when the datasource response is received, only when the conditions of the issue are met:
- The focus is not longer on the selector that was being filtered
- The entered text of the selector is '' (the original text has been deleted)