Fixes issue 21709: Tab title is updated when a record is saved in the grid view
The fix of the issue 21352 caused the updateTabTitle function to be executed when the grid was still in edit mode, so the '*' in the tab title remained. Removing the code that caused this issue was not possible, because then the original issue was reproducible. Instead, an extra call to updateTabTitle is done when the records exists the edit mode.