Fixes issue 21723: Shorcuts work again while editing the grid
If the arrow down, arrow up or enter key was pressed while editing the grid but without any actual changes done, its shortcuts were not being executed. This happened because this keys trigger a validation on the changed fields, and it was not being checked in they were not null.

--HG--
extra : rebase_source : 4169e3ddd7589ec866540712b8724dad49a4f84b