Fixes issue 21716: REST Webservice works with organization '*'
The EntityResolver.getOrgIds has been modified, because prior to the fix if it was called with the organization '0', it returned '0' instead of the natural true. And now, if it is called for an organization other than EntityResolver.organization, it returns a recomputed natural true (orgNaturalTree is the natural the of EntityResolver.organization).

--HG--
extra : rebase_source : eddc11d81144a04aae3de52fc0ebaa624e9cd52c