Fixes issue 21764: Only the proper records are shown after completing an order
The problem was caused by the fix of the issue 21437. Some logic was added to ensure that the recently created records were shown after completing a process, even if the do not comply with the filters. The problem was that this was done on a tab level, without taking into account that their parent record might not be the same.

This has been fixed, and only the child records of the current parent record are shown.