Fixes issue 21741: Pick&Execute under some circumstances is not retrieving the selected record ID
The root cause of this problem is that the context information was getting retrieved from the header tab, instead of the tab that was selected when the pick a
nd execute window was selected. Another change has been done so that that same tab will be selected when the P&E window is closed, otherwise the record could not be refreshed, because the row ID saved in the window belong to the tab that was focused when the P&E window was opened.