Fixes bug 21697: When a date filter is removed, it is applied when tabbing out
Before this fix, in order to delete a date filter the Enter key had to be pressed. This was not consistent, because in order to apply a not-null date filter, it was enough to tab out of the column. Now, if the date filter text is deleted and the column loses the focus, the change will be applied even if the Enter key is not pressed.