Fixes issue 21810: Times are exported to CSV using client timezone
Before this fix, when a hour field was exported, it was done using the timezone of the server. Now the time offset used is the client's side.

In order to do this, when the user pushes the button to export a grid to CSV, the UTC offset of the client is sent to the server.