Fixes issue 20328: The autosave is not done until receiving the FIC response
If an autosave is executed while waiting for a FIC response, the data saved is not the definitive, and might be invalid or incomplete.

When that happens, the actual save will be postponed until the FIC response is received, the same done it is done when the saveRow function is executed.