Fixes issue 21817: SQL comments are applied to the proper columns
There was a problem in postgreSQL that happened when a table was created and one of its columns was removed and created again. The ordinal position of the columns in that case is not consecute, i.e.

Before removing the column:
Ordinal position    Name
1                   Column1
2                   Column2
3                   Column3

After removing the column Column2
Ordinal position    Name
1                   Column1
3                   Column3

After adding the column Column2 again
Ordinal position    Name
1                   Column1
3                   Column3
4                   Column2

When the comments are applied to the columns, instead of using the ordinal position, the index (1..number of columns) was used to determine which comment belongs to each column. This resulted that the comments were applied to the wrong columns. If instead of using the index the ordinal position is used, the problem is solved.

--HG--
extra : rebase_source : 5cd101d21c7e54c125fc6887da73c82b6d2fae98