Fixes issue 20328: Sometimes it is possible to trigger an incomplete autosave
The previous approach used to fix this issue not only did not work in the grid view, but it was not completely correct in the form view. If we allowed to move to the child tab before saving the parent tab, we would be enabling to create a child record, which would fail if the parent record has not been saved yet.

The new approach is to not allow to change the active tab at all until the currently focused tab can be saved (that is, until it has received the FIC response).

--HG--
extra : rebase_source : 04467f02949a48f16836b9e9be7499b190643bd1