Fixes issue 21669: Validation when exporting Database not working properly
There was a validation that was only being done to the tables that belonged to the module being validated, but not to the tables modified by that module. Now, if a module adds a column to a table that belongs to another module but forgets to add it to the application dictionary, the validation will detect the error, the same way it has been working for the columns defined in tables that belonged to the module being validated.