Fixes issue 21910: When a widget is opened in a window it shows all its records
This issue was caused by the fix of the issue 16622. In that issue the ReadOnlyDataSourceServeld was changed so that it kept count of the number of records if the NOCOUNT parameter was set to 'false', as it was done with other datasources. In this case, in order not to break existing calls to the datasource, the count will only be prevented if the parameter is set to 'true', so if the parameter does not have any value (as it is in this scenario of the issue), the count will be done.

--HG--
extra : transplant_source : %EB%EA5A%B5%81%CF%D2%D5%BA%3CVE%AF%8A%1D%B1%C3%C3%E9