Fixes Issue 0021760: First Focused Field is not working in tab Customer
The First Focussed field property is used when creating a new row or while editing in grid. But in few tabs, on tab focus, if the lastFocusedItem is null, it picks up the first field in the form. Modified it take the First Focussed field property into account.

--HG--
extra : rebase_source : 22ba001adc38a0d65b70740ccbd340d7632e66d3