Fixes Issue 0022031: Now selectors are working properly but not as before
The way selectors were working was modified to work properly in this issue https://issues.openbravo.com/view.php?id=21470.Prior to this change, even though 'startsWith' was set, the suggestion box always worked with the 'Contains' filter. Because of the change the behaviour was different from the existing behaviour of the selector, so modified the suggestion text match style from 'startsWith' to 'Contains' so that the existing behaviour will be restored.

--HG--
extra : rebase_source : 10d5638f9b0fddbcd83d381e907e668bdbbfbf64