Fixes issue 22438: Display logic works fine for new records
The problem was caused a by change done in [1], where some call to this.setValue were changed to this.setItemValue. The setItemValue function was not updating the value if the original value of the field was undefined and the new value was null.

[1] https://code.openbravo.com/erp/devel/pi/rev/f786d405627ea5465bb6a330af5aed800790ca6d

--HG--
extra : transplant_source : %03%AEj/%5E%A3%11LBY%0C%C6%3BP%A0%D8i8%86%EE